+++ bro_abstract_manager/lib/bro_abstract_manager.dart
@@ -4,5 +4,9 @@
 
 library;
 
-export 'src/abs_manager_builder.dart';
-export 'src/abs_with_life_cycle.dart';
+export 'package:bro_platform_utility/bro_platform_utility.dart' show PlatformType;
+
+export 'src/errors/platform_not_supported_by_manager_error.dart';
+export 'src/services/abs_manager_builder.dart';
+export 'src/services/abs_with_life_cycle.dart';
+export 'src/types/wrong_platform_behavior.dart';