+++ bro_firebase_crashlytics/lib/bro_firebase_crashlytics.dart
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2025 Benoit Rolandeau <borlnov.obsessio@gmail.com>
+//
+// SPDX-License-Identifier: MIT
+
+library;
+
+export 'src/mixins/mixin_crashlytics_configs.dart';
+export 'src/services/crashlytics_logger_manager.dart';