+++ bro_abstract_manager/CHANGELOG.md
@@ -7,6 +7,13 @@ SPDX-License-Identifier: MIT
 <!-- We disable the rule for having the top level at first line, because we prefer a second level
      header for version -->
 <!-- markdownlint-disable MD041 -->
+## 1.1.0
+
+- Add new methods to the builder classes to know if the manager supports (or not) the current
+  platform.
+- Transform the async factory passed to the builder class constructor into a method to override by
+  derived class.
+
 ## 1.0.1
 
 - Update pubspec.yaml file to be more compliant with dart pub.dev requirements.